requirements.txt
streamlit
pinecone
openai
PyPDF2
python-docx
python-dotenv